+++ CHANGES.md
@@ -1,6 +1,8 @@
 
 * Add `Ptime.v` and `Ptime.Span.v` to safely deal with trusted
   inputs. Thanks to Matt Gray for suggesting.
+* Add `Ptime.weekday`, to help conversions to denormalized
+  timestamp formats. Thanks to Romain Calascibetta for suggesting.
 * Build depend on topkg.
 * Relicense from BSD3 to ISC.